README.md
This is the start of a new journey for me. I plan to learn as much as I can as fast as I can. I’m kind of putting myself trough a self-imposed boot camp on coding/programming. I am open for suggestions, advice,  and guidance as I grow into the developer that can and will colllaborate on projects with other fellow developers.